SQL select 분리
1. select문은 어느 table에서 사용해도 같은 기능만 하기 때문에 util > sql로 뺌
